Aggregated enrichers in their own features
This allow us to turn off parts of SC easier
