+++ src/ServiceControl/ServiceControl.csproj
@@ -239,7 +239,7 @@
     <Compile Include="Contracts\MessageFailures\MessageSubmittedForRetry.cs" />
     <Compile Include="Contracts\Operations\EndpointDetails.cs" />
     <Compile Include="Contracts\EndpointControl\MonitoringDisabledForEndpoint.cs" />
-    <Compile Include="Contracts\Operations\EnrichWithTrackingIds.cs" />
+    <Compile Include="Contracts\Operations\DefaultEnrichers.cs" />
     <Compile Include="Contracts\Operations\ExceptionDetails.cs" />
     <Compile Include="Contracts\Operations\FailureDetails.cs" />
     <Compile Include="CustomChecks\CustomCheckNotifications.cs" />
@@ -313,7 +313,6 @@
     <Compile Include="CustomChecks\DeleteCustomCheck.cs" />
     <Compile Include="CustomChecks\DeleteCustomCheckHandler.cs" />
     <Compile Include="EndpointControl\DetectNewEndpointsFromImportsEnricher.cs" />
-    <Compile Include="EndpointControl\EnrichWithEndpointDetails.cs" />
     <Compile Include="EndpointControl\Handlers\DetectEndpointsFromHeartbeats.cs" />
     <Compile Include="EndpointControl\Handlers\KnownEndpointsCache.cs" />
     <Compile Include="EndpointControl\Handlers\RegisterEndpointHandler.cs" />
@@ -383,7 +382,6 @@
     <Compile Include="MessageFailures\Handlers\ArchiveMessageHandler.cs" />
     <Compile Include="MessageFailures\Handlers\ImportFailedMessageHandler.cs" />
     <Compile Include="MessageFailures\FailedMessage.cs" />
-    <Compile Include="MessageFailures\DetectSuccessfullRetriesEnricher.cs" />
     <Compile Include="MessageFailures\Handlers\MessageFailureResolvedHandler.cs" />
     <Compile Include="MessageFailures\InternalMessages\ArchiveMessage.cs" />
     <Compile Include="MessageFailures\InternalMessages\PerformRetry.cs" />
@@ -408,7 +406,6 @@
     <Compile Include="Infrastructure\OWIN\Startup.cs" />
     <Compile Include="Infrastructure\Settings\Settings.cs" />
     <Compile Include="Infrastructure\Settings\SettingsReader.cs" />
-    <Compile Include="MessageTypes\MessageTypeEnricher.cs" />
     <Compile Include="Operations\AuditQueueImport.cs" />
     <Compile Include="CompositeViews\Messages\GetMessagesByConversation.cs" />
     <Compile Include="Bootstrapper.cs" />
@@ -420,7 +417,6 @@
     <Compile Include="Operations\BodyStorage\BodyStorageFeature.cs" />
     <Compile Include="Operations\BodyStorage\RavenAttachments\GetBodyById.cs" />
     <Compile Include="Operations\BodyStorage\IBodyStorage.cs" />
-    <Compile Include="Operations\BodyStorage\BodyStorageEnricher.cs" />
     <Compile Include="Operations\BodyStorage\RavenAttachments\RavenAttachmentsBodyStorage.cs" />
     <Compile Include="Operations\ErrorQueueImport.cs" />
     <Compile Include="Operations\ExceptionExtensions.cs" />
@@ -432,13 +428,12 @@
     <Compile Include="Operations\MsmqAuditImporterPerformanceCounters.cs" />
     <Compile Include="Operations\SatelliteImportFailuresHandler.cs" />
     <Compile Include="Operations\ImportFailureCircuitBreaker.cs" />
-    <Compile Include="Operations\ImportRegistration.cs" />
     <Compile Include="Infrastructure\TypeExtensions.cs" />
-    <Compile Include="Operations\UpdateLicenseEnricher.cs" />
+    <Compile Include="Operations\LicenseReporter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Infrastructure\RavenDB\RavenBootstrapper.cs" />
-    <Compile Include="MessageAuditing\ProcessingStatisticsEnricher.cs" />
+    <Compile Include="MessageAuditing\ProcessingStatistics.cs" />
     <Compile Include="Operations\DetectFailedMessageImportsFeature.cs" />
     <Compile Include="Recoverability\Grouping\Archiving\ArchiveAllInGroup.cs" />
     <Compile Include="Recoverability\Grouping\Archiving\ArchiveAllInGroupHandler.cs" />
@@ -478,7 +473,7 @@
     <Compile Include="SagaAudit\Messages\SagaChangeInitiator.cs" />
     <Compile Include="SagaAudit\Messages\SagaChangeOutput.cs" />
     <Compile Include="SagaAudit\Messages\SagaUpdatedMessage.cs" />
-    <Compile Include="SagaAudit\SagaRelationshipsEnricher.cs" />
+    <Compile Include="SagaAudit\SagaAuditing.cs" />
     <Compile Include="SagaAudit\ProcessingState.cs" />
     <Compile Include="SagaAudit\SagaInfo.cs" />
     <Compile Include="SagaAudit\SagaUpdatedHandler.cs" />
